swapwing_backend/requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
Pillow==9.5.0; python_version < "3.12"
Pillow>=10.1,<11; python_version >= "3.12"
pytz==2023.3
sqlparse==0.4.3
psycopg2>=2.8
redis>=3.5
celery>=5.0
channels-redis
requests
pyfcm
drf-spectacular>=0.26,<0.27
gunicorn>=20.1
django-storages[boto3]>=1.13
boto3>=1.34
pytest>=7.4,<8
pytest-django>=4.5,<4.6
pytest-cov>=4.1,<5
coverage>=7.3,<8
